pyproject.toml
[project]
name = "process-ave"
version = "0.1.0"
description = "Open source checklist templates and runs API"
authors = [{ name = "Process Ave" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.1",
    "SQLAlchemy>=2.0.27",
    "psycopg[binary,pool]>=3.1.18",
    "pydantic>=2.6.3",
    "alembic>=1.13.1",
    "python-dotenv>=1.0.1"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.1",
    "httpx>=0.27.0"
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.uvicorn]
factory = "app.main:create_app"

